requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
pydantic==2.5.0
